+++ isis/src/base/objs/Strategy/Strategy.h
@@ -15,6 +15,7 @@ find files of those names at the top level of this repository. **/
 #include <QString>
 
 // SharedResource and ResourceList typedefs
+#include "GisGeometry.h"
 #include "Resource.h"
 #include "PvlObject.h"
 
@@ -55,7 +56,7 @@ namespace Isis {
    * @internal 
    *   @history 2012-07-15 Kris Becker - Original version.
    *   @history 2015-03-18 Jeannie Backer - Brought class files closer to ISIS coding standards.
-   *   @history 2015-04-07 Kristin Berry - Modified applytoIntserectedGeometry to deal
+   *   @history 2015-04-07 Kristin Berry - Modified applytoIntersectedGeometry to deal
    *   @history 2015-03-26 Jeannie Backer - Updated documentation.
    *   @history 2015-03-28 Kris Becker - Added the composite(Resource, Resource)
    *                          method
@@ -80,6 +81,12 @@ namespace Isis {
    *                          replacements in reverse order fixes this issue.
    * @history 2016-03-07  Tyler Wilson - Corrected documentation, and created a
    *                          unit test to test most of this classe's methods.
+   * @history 2018-07-29 Kris Becker - Fixed errors in importGeometry() where an 
+   *                          exception was being thrown when checking for valid
+   *                          Geometry. It now checks for this case and runs
+   *                          buffer(0) to attempt to fix the geometry. This is
+   *                          typically a self-intersection error that can be
+   *                          repaired with a buffer(0) operator.
    */
 
   class Strategy {