+++ CHANGELOG.md
@@ -45,6 +45,7 @@ release.
 - Updated pixel2map documentation
 
 ### Fixed
+- Fixed a bug in isisminer in which bad (e.g. self-intersecting) polygon geometries were not treated properly. Added pertinent unit tests to GisGeometry and Strategy classes. Issue: [5612](https://github.com/DOI-USGS/ISIS3/issues/5612)
 - Fixed a bug in kaguyasp2isis that doesn't work for data with a detached label.
 
 ## [8.3.0] - 2024-09-30